feat: 0.12.2 — promote Events to Alloy.Events (correction of 0.12.1 deprecation)
The 0.12.1 release deprecated Alloy.Agent.Events in favor of
AlloyAgent.Events, treating event envelope construction as a runtime
concern. That classification was wrong: Alloy.Agent.Turn calls Events
directly, so the v1 envelope format is part of Alloy's protocol surface
and stays in Alloy.

- New canonical module Alloy.Events at lib/alloy/events.ex
- Alloy.Agent.Events becomes a @deprecated thin shim that delegates
  to Alloy.Events (still removed in 0.13.0, only the destination changed)
- Alloy.Agent.Turn updated to alias Alloy.Events
- CHANGELOG and docs/MIGRATION-alloy_agent.md retract the 0.12.1
  guidance to migrate Events to alloy_agent
- Server and Session deprecations are unchanged — those modules really
  are runtime concerns Turn doesn't depend on
